+++ spacemake/__init__.py
@@ -1,1 +1,4 @@
 __version__ = 1.0
+
+from . import preprocess as pp
+from . import spatial as sp
