+++ spacemake/preprocess/__init__.py
@@ -1,0 +1,5 @@
+from ._cmdline import cmdline
+from ._adata import calculate_adata_metrics,\
+    calculate_shannon_entropy_scompression, dge_to_sparse_adata,\
+    attach_barcode_file, parse_barcode_file
+
