+++ spacemake/preprocess/_fastq.py
@@ -1063,29 +1063,3 @@ def parse_args():
     args = parser.parse_args()
 
     return args
-
-
-def cmdline():
-    with ExceptionLogging("main"):
-        args = parse_args()
-        NO_CALL = args.na
-        setup_logging(args)
-
-        if args.out_format == "bam" and not args.read2:
-            raise ValueError("bam output format requires --read2 parameter")
-
-        if ("bc1" in args.cell and not args.bc1_ref) or (
-            "bc2" in args.cell and not args.bc2_ref
-        ):
-            raise ValueError(
-                "bc1/2 are referenced in --cell or --cell-raw, but no reference barcodes are specified via --bc{{1,2}}-ref"
-            )
-
-        if args.bc1_ref or args.bc2_ref:
-            main_combinatorial(args)
-        else:
-            main_dropseq(args)
-
-
-if __name__ == "__main__":
-    cmdline()
